requirements.txt
opencv-python >= 4.8.0.74
pytesseract >= 0.3.10
requests >= 2.25.1
Pillow >= 9.5.0
loguru>=0.6.0